+++ Snake/src/main/java/handler/BaseHandler.java
@@ -1,5 +1,6 @@
+package handler;
+import request.Request;
 
-import java.util.*;
 
 /**
  *
