+++ Snake/src/main/java/handler/Handler.java
@@ -1,5 +1,6 @@
+package handler;
 
-import java.util.*;
+import request.Request;
 
 /**
  *
