+++ Snake/src/main/java/handler/HandlerFactory.java
@@ -1,5 +1,4 @@
-
-import java.util.*;
+package handler;
 
 /**
  *
