+++ Snake/src/main/java/handler/SnakeHandler/ColorSegment.java
@@ -1,6 +1,6 @@
+package handler.SnakeHandler;
 
 import java.awt.*;
-import java.util.*;
 
 /**
  *
