+++ Snake/src/main/java/handler/SnakeHandler/Segment.java
@@ -1,5 +1,4 @@
-
-import java.util.*;
+package handler.SnakeHandler;
 
 /**
  *
