+++ Snake/src/main/java/request/Request.java
@@ -1,2 +1,6 @@
+package request;
+
 public interface Request {
+
+
 }
